services/llgen/requirements.txt
click
scapy-python3
